+++ fakturogen-invoices/src/main/java/pl/fakturogen/fakturogeninvoices/model/entities/Customer.java
@@ -1,4 +1,4 @@
-package pl.fakturogen.fakturogeninvoices.model;
+package pl.fakturogen.fakturogeninvoices.model.entities;
 
 public class Customer {
 }
